% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.sclr}
\alias{coef.sclr}
\alias{vcov.sclr}
\title{ML estimate components}
\usage{
\method{coef}{sclr}(object, ...)

\method{vcov}{sclr}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{sclr}}.}

\item{...}{Not used. Needed to match generic signature.}
}
\description{
\code{coef} returns MLE's.
\code{vcov} returns the estimated variance-covariance matrix at MLE's.
}
