% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protection.R
\name{get_protection_data}
\alias{get_protection_data}
\title{Protection curve data}
\usage{
get_protection_data(fit, ci_level = 0.95, log_titre_range = c(0, 8),
  n_points = 101)
}
\arguments{
\item{fit}{Object returned by \code{\link{sclr}}.}

\item{ci_level}{Confidence level for the curve.}

\item{log_titre_range}{Range of log titres over which to calculate the 
protection.}

\item{n_points}{Amount of points in \code{log_titre_range}.}
}
\value{
A dataframe.
}
\description{
Calculates the data required for the protection curve.
}
