% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protection.R
\name{get_protection_level}
\alias{get_protection_level}
\title{Protection level calculations}
\usage{
get_protection_level(fit, newdata, var_name, lvl = 0.5,
  ci_level = 0.95, tol = 10^(-7))
}
\arguments{
\item{fit}{Object returned by \code{\link{sclr}}.}

\item{newdata}{A dataframe with all covariates except the one for which
protection values should be calculated.}

\item{var_name}{Name of the covariate for which the protection values should
be calculated. This name should appear in the formula of the call to
\code{\link{sclr}} which was used to generate \code{fit}.}

\item{lvl}{Protection level to find titre values for. Default is 0.5 (50\%)}

\item{ci_level}{Confidence level for the calculated interval. 
Default is 0.95.}

\item{tol}{Tolerance. The values will be found numerically,
once the algorithm converges within \code{tol} of \code{lvl} 
it stops looking.}
}
\value{
A dataframe. Will have the same variables as \code{newdata} with
the addition of the \code{var_name} variable.
}
\description{
Calculates covariate values corresponding to a particular protection level.
Only accepts one covariate at a time, fixed 
values of all the others should be provided.
}
