% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.sclr}
\alias{predict.sclr}
\title{Predict method for scaled logit model fit.}
\usage{
\method{predict}{sclr}(fit, newdata, ci_lvl = 0.95)
}
\arguments{
\item{fit}{}

\item{newdata}{}

\item{ci_lvl}{}
}
\description{
Returns only the protection estimates.
}
