% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sclr}
\alias{print.sclr}
\alias{summary.sclr}
\title{Print a \code{sclr} object.}
\usage{
\method{print}{sclr}(fit)

\method{summary}{sclr}(fit)
}
\arguments{
\item{fit}{An object returned by \code{\link{sclr}}.}
}
\description{
Summarises a \code{sclr} object for printing. For a dataframe summary, see
\code{\link[=tidy.sclr]{tidy}}.
}
