% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{sclr}
\alias{sclr}
\title{Fits the scaled logit model}
\usage{
sclr(formula, data, tol = 10^(-7), n_iter = NULL, max_tol_it = 10^4)
}
\arguments{
\item{formula}{an object of class "formula":
a symbolic description of the model to be fitted.}

\item{data}{a data frame.}

\item{tol}{Tolerance. Used when \code{n_iter} is \code{NULL}.}

\item{n_iter}{Number of Newton-Raphson iterations. \code{tol} is ignored
when this is not \code{NULL}.}

\item{max_tol_it}{Maximum tolerated iterations. If it fails to coverge 
within this number of iterations, will return with an error.}
}
\description{
Used to fit the scaled logit model from Dunning (2006).
}
