% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{sclr_log_likelihood}
\alias{sclr_log_likelihood}
\title{Log-likelihood}
\usage{
sclr_log_likelihood(fit, pars = NULL)
}
\arguments{
\item{fit}{An object returned by \code{\link{sclr}}. Or a list with
parameters, x and y entries corresponding to the parameter matrix, 
model matrix and model response.}

\item{pars}{A named vector of parameter values. If \code{NULL} then the
estimates from \code{fit} will be used.}
}
\description{
Computes the log-likelihood of the scaled logit model at a given set of
parameter estimates (or the MLE if \code{pars} is not supplied).
}
