% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{testing_documentation}
\alias{testing_documentation}
\title{Temporary funciton for testing purposes}
\usage{
testing_documentation(x)
}
\arguments{
\item{x}{Anything that can be passed to cat. New bit.}
}
\description{
Temporary funciton for testing purposes
}
\examples{
testing_documentation("test")
}
