% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.sclr}
\alias{tidy.sclr}
\title{Summarise \code{sclr} fit in a tibble}
\usage{
tidy.sclr(fit, ci_level = 0.95)
}
\arguments{
\item{fit}{An object returned by \code{sclr}.}

\item{ci_level}{Confidence level for the intervals}
}
\value{
A tibble.
}
\description{
Summarise \code{sclr} fit in a tibble
}
