% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.sclr}
\alias{vcov.sclr}
\title{Variance-covariance matrix}
\usage{
\method{vcov}{sclr}(fit)
}
\arguments{
\item{fit}{An object returned by \code{\link{sclr}}.}
}
\description{
Returns the estimated variance-covariance matrix.
}
