% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.sclr}
\alias{coef.sclr}
\title{Coefficients}
\usage{
\method{coef}{sclr}(fit)
}
\arguments{
\item{fit}{An object returned by \code{\link{sclr}}.}
}
\description{
Returns the estimated model coefficients.
}
