% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protection.R
\name{get_protection_level}
\alias{get_protection_level}
\title{Protection level titres.}
\usage{
get_protection_level(fit, name, lvl = 0.5, ci_level = 0.95,
  tol = 10^(-7))
}
\arguments{
\item{fit}{Object returned by \code{\link{sclr}}.}

\item{lvl}{Protection level to find titre values for.}

\item{ci_level}{Confidence level.}

\item{tol}{Tolerance.}
}
\value{
Named vector.
}
\description{
Calculates titres corresponding to a particular protection level.
}
