% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sclr}
\alias{print.sclr}
\alias{summary.sclr}
\title{Print a \code{sclr} object.}
\usage{
\method{print}{sclr}(x, ...)

\method{summary}{sclr}(object, ...)
}
\arguments{
\item{x, object}{An object returned by \code{\link{sclr}}.}

\item{...}{Not used. Needed to match generic signature.}
}
\description{
Summarises a \code{sclr} object for printing. For a dataframe summary, see
\code{\link[=tidy.sclr]{tidy}}.
}
