% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{sclr}
\alias{sclr}
\title{Fits the scaled logit model}
\usage{
sclr(formula, data, calc_ci = TRUE, ci_lvl = 0.95, calc_ll = TRUE,
  tol = 10^(-7), n_iter = NULL, max_tol_it = 10^4)
}
\arguments{
\item{formula}{an object of class "formula":
a symbolic description of the model to be fitted.}

\item{data}{a data frame.}

\item{calc_ci}{Whether to calculate confindence intervals. 
\code{TRUE} by default.}

\item{ci_lvl}{Confidence interval level for the parameter estimates.}

\item{calc_ll}{Whether to calculate log likelihood at MLEs. 
\code{TRUE} by default.}

\item{tol}{Tolerance. Used when \code{n_iter} is \code{NULL}.}

\item{n_iter}{Number of Newton-Raphson iterations. \code{tol} is ignored
when this is not \code{NULL}.}

\item{max_tol_it}{Maximum tolerated iterations. If it fails to coverge 
within this number of iterations, will return with an error.}
}
\value{
An object of class \code{sclr}. This is a list with the following
elements:
\describe{
  \item{paramters}{Maximum likelihood estimates of the parameter values.}
  \item{covariance_mat}{The variance-convariance matrix of the parameter
    estimates.}
  \item{n_converge}{The number of Newton-Raphson iterations 
    (including resets) that were required for convergence.}
  \item{confint}{Confidence intervals of the parameter estimates.}
  \item{x}{Model matrix derived from \code{formula} and \code{data}.}
  \item{y}{Response matrix derived from \code{formula} and \code{data}.}
  \item{log_likelihood}{Value of log-likelihood calculated at the ML 
    estimates of parameters.}
  \item{call}{The original call to \code{sclr}.}
  \item{model}{Model frame object derived from 
    \code{formula} and \code{data}.}
  \item{terms}{Terms object derived from model frame.}
}
Methods supported: \code{print}, \code{vcov}, \code{coef}, 
\code{summary}, \code{predict}, \code{tidy} (\code{broom} package)
}
\description{
Used to fit the scaled logit model from Dunning (2006).
}
\details{
The model is of the form 
\deqn{P(Y = 1) = \lambda(1 - logit^{-1}(\beta_0 + \beta_1X_1 + 
  \beta_2X_2 + ... + \beta_kX_k))}
Where \eqn{Y} is the binary outcome idicator, 
(eg. 1 - infected, 0 - not infected). \eqn{X} - covariate. 
\eqn{k} - number of covariates.
The likelihood maximisation uses the Newton-Raphson algorithm. 
Initial values are always 1 for the covariate coefficients 
(and the associated intercept) and the proportion of infected for the 
baseline risk. If an algorithm's iteration produces values that cannot 
be used, a different set of initial values is chosen randomly and the 
algorithm starts over. Unusable parameter values are: baseline risk outside
of (0, 1) (likelihood undefined) and covariate coefficients (or the 
associated intercept) greater than 100 in magnitude (these terms 
appear in the exponent, if they are too big then calculations will fail).
}
\references{
Dunning AJ (2006). 
"A model for immunological correlates of protection." 
Statistics in Medicine, 25(9), 1485-1497. doi: 10.1002/sim.2282.
}
