% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{sclr_fit}
\alias{sclr_fit}
\title{Fitter function for the scaled logit model}
\usage{
sclr_fit(y, x, tol = 10^(-7), n_iter = NULL)
}
\arguments{
\item{y}{A vector of observations.}

\item{x}{A design matrix.}

\item{tol}{Tolerance. Used when \code{n_iter} is \code{NULL}.}

\item{n_iter}{Number of Newton-Raphson iterations. \code{tol} is ignored
when this is not \code{NULL}.}
}
\description{
Computing engine behind \code{\link{sclr}}.
}
