% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{sclr_log_likelihood}
\alias{sclr_log_likelihood}
\title{Log-likelihood}
\usage{
sclr_log_likelihood(y, x, pars)
}
\arguments{
\item{y}{A vector of observations.}

\item{x}{A design matrix.}

\item{pars}{A named vector of parameter values.}
}
\description{
Computes the log-likelihood of the scaled logit model
at a given set of parameter values.
}
